Move parallel note commitment trees to zebra-chain
