+++ zebra-chain/Cargo.toml
@@ -81,6 +81,7 @@ rand_chacha = { version = "0.3.1", optional = true }
 tokio = { version = "1.20.0", features = ["tracing"], optional = true }
 
 zebra-test = { path = "../zebra-test/", optional = true }
+rayon = "1.5.3"
 
 [dev-dependencies]
