+++ zebra-chain/src/parallel.rs
@@ -1,0 +1,3 @@
+//! Parallel chain update methods.
+
+pub mod tree;
