+++ zebra-state/src/service/finalized_state/zebra_db/block.rs
@@ -21,6 +21,7 @@ use zebra_chain::{
     block::{self, Block, Height},
     history_tree::HistoryTree,
     orchard,
+    parallel::tree::NoteCommitmentTrees,
     parameters::{Network, GENESIS_PREVIOUS_BLOCK_HASH},
     sapling,
     serialization::TrustedPreallocate,
@@ -36,7 +37,7 @@ use crate::{
             block::TransactionLocation,
             transparent::{AddressBalanceLocation, OutputLocation},
         },
-        zebra_db::{metrics::block_precommit_metrics, shielded::NoteCommitmentTrees, ZebraDb},
+        zebra_db::{metrics::block_precommit_metrics, ZebraDb},
         FinalizedBlock,
     },
     BoxError, HashOrHeight,
